DocAPI V2.1 and Some features of quiz

+++ src/main/java/com/quizztoast/backendAPI/repository/QuizRepository.java
@@ -4,5 +4,7 @@ import com.quizztoast.backendAPI.model.quiz.Quiz;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface QuizRepository extends JpaRepository<Quiz,Integer> {
+public interface QuizRepository extends JpaRepository<Quiz,Long> {
+
+
 }
